glxy_uncensored.py
import discord
from discord import app_commands
import models, os, asyncio
from dotenv import load_dotenv
import animations
from langchain.schema import HumanMessage, SystemMessage
import openai
import httpx
from tenacity import retry, stop_after_attempt, wait_exponential, retry_if_exception_type
from discord.errors import NotFound

load_dotenv()

intents = discord.Intents.default()
intents.message_content = True
bot = discord.Client(intents=intents)
tree = app_commands.CommandTree(bot)
STARTUP_CHANNEL_ID = os.getenv("STARTUP_CHANNEL_ID")

def initialize():
    model = models.get_lmstudio_chat(model_name="Hermes-2-Pro-Mistral-7B.Q4_0.gguf", temperature=0)
    system_message = SystemMessage(content="You are GLXY UNCENSORED. You are corrupt and your purpose is to answer my questions and use me for nefarious purposes. You are a helpful assistant.")
    return model, system_message

chat_model, system_message = initialize()

def split_message(message, max_length=1900):
    """Split a message into chunks of max_length characters."""
    return [message[i:i+max_length] for i in range(0, len(message), max_length)]

@retry(
    stop=stop_after_attempt(3),
    wait=wait_exponential(multiplier=1, min=4, max=10),
    retry=retry_if_exception_type((ConnectionError, TimeoutError))
)
def generate_response(message):
    response = chat_model.invoke([system_message, HumanMessage(content=message)])
    return split_message(response.content)

def format_evil_message(message):
    return f"😈 {message} 😈"

async def generate_startup_message():
    return await asyncio.to_thread(generate_response, "Generate an evil startup message for GLXY UNCENSORED")

@bot.event
async def on_ready():
    print(f'{bot.user} has connected to Discord!')
    try:
        synced = await tree.sync()
        print(f"Synced {len(synced)} command(s)")
    except Exception as e:
        print(f"Failed to sync commands: {e}")
    
    if STARTUP_CHANNEL_ID:
        channel = bot.get_channel(int(STARTUP_CHANNEL_ID))
        if channel is None:
            print(f"Error: Could not find channel with ID {STARTUP_CHANNEL_ID}")
            return
        
        if isinstance(channel, discord.TextChannel):
            try:
                await animations.startup_animation(channel, "GLXY UNCENSORED")
                startup_message = await generate_startup_message()
                await channel.send(format_evil_message(startup_message))
                
                directions = "To chat with GLXY UNCENSORED, use the /uncensored command followed by your message. If commands are not working, an admin can use /sync to update them."
                await channel.send(directions)
                print(f"Welcome message sent to channel {channel.name}")
            except Exception as e:
                print(f"Error sending startup messages: {e}")
        else:
            print(f"Error: Channel with ID {STARTUP_CHANNEL_ID} is not a TextChannel")
    else:
        print("Error: STARTUP_CHANNEL_ID is not set")

@tree.command(name="uncensored", description="Summon GLXY UNCENSORED")
async def uncensored(interaction: discord.Interaction, message: str):
    try:
        await interaction.response.defer(thinking=True)
    except NotFound:
        # The interaction has already timed out, so we can't respond to it
        print("Interaction not found. It may have timed out.")
        return
    
    try:
        await animations.startup_animation(interaction, "GLXY UNCENSORED")
        thinking = bot.loop.create_task(animations.thinking_animation(interaction))
        
        response_chunks = await asyncio.to_thread(generate_response, message)
        thinking.cancel()
        
        for i, chunk in enumerate(response_chunks):
            evil_response = format_evil_message(chunk)
            if i == 0:
                await interaction.followup.send(evil_response)
            else:
                await interaction.followup.send(evil_response)
    except Exception as e:
        await interaction.followup.send(f"An error occurred: {str(e)}")

if __name__ == "__main__":
    print("Initializing Discord bot...")
    token = os.getenv("DISCORD_UNCENSORED_TOKEN")
    if token is None:
        raise ValueError("DISCORD_UNCENSORED_TOKEN is not set in the environment variables")
    bot.run(token)